moved model to its own module
